refactor(connections): split connections_panel into part files
Extract sidebar, Redis, Mongo, Postgres connection tile, MySQL, and PG tree
widgets into part libraries. Behavior unchanged; easier navigation and review.
